@@ -4,7 +4,7 @@ version: 1.5.0
 homepage: https://github.com/tomalabaster/flutter_app_lock
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.12.1"
 
 dependencies:
