@@ -22,22 +22,22 @@ import 'package:flutter/material.dart';
 /// the app is in the background state before the [lockScreen] widget should be
 /// shown upon returning. It defaults to instantly.
 class AppLock extends StatefulWidget {
-  final Widget Function(Object) builder;
+  final Widget Function(Object?) builder;
   final Widget lockScreen;
   final bool enabled;
   final Duration backgroundLockLatency;
-  final ThemeData theme;
+  final ThemeData? theme;
 
   const AppLock({
-    Key key,
-    @required this.builder,
-    @required this.lockScreen,
+    Key? key,
+    required this.builder,
+    required this.lockScreen,
     this.enabled = true,
     this.backgroundLockLatency = const Duration(seconds: 0),
     this.theme,
   }) : super(key: key);
 
-  static _AppLockState of(BuildContext context) =>
+  static _AppLockState? of(BuildContext context) =>
       context.findAncestorStateOfType<_AppLockState>();
 
   @override
@@ -47,15 +47,15 @@ class AppLock extends StatefulWidget {
 class _AppLockState extends State<AppLock> with WidgetsBindingObserver {
   static final GlobalKey<NavigatorState> _navigatorKey = GlobalKey();
 
-  bool _didUnlockForAppLaunch;
-  bool _isLocked;
-  bool _enabled;
+  late bool _didUnlockForAppLaunch;
+  late bool _isLocked;
+  late bool _enabled;
 
-  Timer _backgroundLockLatencyTimer;
+  Timer? _backgroundLockLatencyTimer;
 
   @override
   void initState() {
-    WidgetsBinding.instance.addObserver(this);
+    WidgetsBinding.instance!.addObserver(this);
 
     this._didUnlockForAppLaunch = !this.widget.enabled;
     this._isLocked = false;
@@ -85,7 +85,7 @@ class _AppLockState extends State<AppLock> with WidgetsBindingObserver {
 
   @override
   void dispose() {
-    WidgetsBinding.instance.removeObserver(this);
+    WidgetsBinding.instance!.removeObserver(this);
 
     this._backgroundLockLatencyTimer?.cancel();
 
@@ -101,7 +101,7 @@ class _AppLockState extends State<AppLock> with WidgetsBindingObserver {
       routes: {
         '/lock-screen': (context) => this._lockScreen,
         '/unlocked': (context) =>
-            this.widget.builder(ModalRoute.of(context).settings.arguments)
+            this.widget.builder(ModalRoute.of(context)!.settings.arguments)
       },
       theme: this.widget.theme,
     );
@@ -122,7 +122,7 @@ class _AppLockState extends State<AppLock> with WidgetsBindingObserver {
   /// when built. Use this when you want to inject objects created from the
   /// [lockScreen] in to the rest of your app so you can better guarantee that some
   /// objects, services or databases are already instantiated before using them.
-  void didUnlock([Object args]) {
+  void didUnlock([Object? args]) {
     if (this._didUnlockForAppLaunch) {
       this._didUnlockOnAppPaused();
     } else {
@@ -161,17 +161,17 @@ class _AppLockState extends State<AppLock> with WidgetsBindingObserver {
   /// Manually show the [lockScreen].
   Future<void> showLockScreen() {
     this._isLocked = true;
-    return _navigatorKey.currentState.pushNamed('/lock-screen');
+    return _navigatorKey.currentState!.pushNamed('/lock-screen');
   }
 
-  void _didUnlockOnAppLaunch(Object args) {
+  void _didUnlockOnAppLaunch(Object? args) {
     this._didUnlockForAppLaunch = true;
-    _navigatorKey.currentState
+    _navigatorKey.currentState!
         .pushReplacementNamed('/unlocked', arguments: args);
   }
 
   void _didUnlockOnAppPaused() {
     this._isLocked = false;
-    _navigatorKey.currentState.pop();
+    _navigatorKey.currentState!.pop();
   }
 }
